+++ CHANGELOG.md
@@ -26,6 +26,12 @@
 
 Unreleased changes
 --------------
+* **Added**
+  * Added arg, `--prt_study_gos_only`, to script, `scripts/find_enrichment.py`
+    to print only study GOs when printing all GO terms, regardless of their significance (`--pval=1.0`):
+    `find_enrichment.py study_genes.txt human_genes.txt gene2go --pval=1.0 --prt_study_gos_only`
+* **Changed**
+  * Remove trailing divider ("NOT|"), if it exists in the gpad file ([go-annotation #2885](https://github.com/geneontology/go-annotation/issues/2885))
 
 Release 2020-03-13 1.0.3
 -------------------------
@@ -181,5 +187,5 @@ Release 2010-04-12
 Release 2010-02-23
 -------------------------
 
-
+https://keepachangelog.com/en/1.0.0/
 git log --since=2019-05-08 --before=2019-07-26